agent/src/requirements.txt
psycopg2==2.7.5
psutil==5.4.7
requests==2.9.1
python-dotenv==0.9.1
json2xml==2.2.1
